Better Remove from Cart
I decided to use a different method to add/remove items from the cart using only one State.

Because each product has only 1 id, but 3 different sizes, it was easy to add them to the cart, but not to remove them.

Rather than the conventional way of removing an item by ID or even by size, I utilized the indexOf() method to pinpoint the exact item, and then remove it through the .filter() method.